Code Review

+++ src/me/crazyrealms/crazyenchants/listeners/BlockBreak.java
@@ -15,7 +15,7 @@ public class BlockBreak implements Listener {
         Player p = e.getPlayer();
         if (Enchant.getEnchants(p.getItemInHand(), p.getInventory().getBoots(), p.getInventory().getChestplate(), p.getInventory().getHelmet(), p.getInventory().getLeggings()) != null) {
             for (Enchant ench : Enchant.getEnchantsOnPlayer(p).keySet()) {
-                int chance = (int) (ench.getChance() + 5 * (Enchant.getEnchantsOnPlayer(p).get(ench)-1));
+                int chance = ench.getChance() + 5 * (Enchant.getEnchantsOnPlayer(p).get(ench)-1);
                 if (chance > new Random().nextInt(100))
                     ench.playerBreakBlockEvent(e);
             }
