+++ src/me/crazyrealms/crazyenchants/listeners/EntityDamage.java
@@ -19,7 +19,7 @@ public class EntityDamage implements Listener {
         Player p = (Player) e.getDamager();
         if(Enchant.getEnchantsOnPlayer(p) != null) {
             for(Enchant ench : Enchant.getEnchantsOnPlayer(p).keySet()) {
-                int chance = (int) (ench.getChance() + 5 * (Enchant.getEnchantsOnPlayer(p).get(ench)-1));
+                int chance = ench.getChance() + 5 * (Enchant.getEnchantsOnPlayer(p).get(ench)-1);
                 if(chance > new Random().nextInt(100)) {
                     ench.playerHitEvent(e);
                 }
@@ -35,7 +35,7 @@ public class EntityDamage implements Listener {
         Player p = (Player) e.getEntity();
         if(Enchant.getEnchantsOnPlayer(p) != null) {
             for(Enchant ench : Enchant.getEnchantsOnPlayer(p).keySet()) {
-                int chance = (int) (ench.getChance() + 5 * (Enchant.getEnchantsOnPlayer(p).get(ench)-1));
+                int chance = ench.getChance() + 5 * (Enchant.getEnchantsOnPlayer(p).get(ench)-1);
                 if(chance > new Random().nextInt(100)) {
                     ench.genericDamageEvent(e);
                 }
